+++ packages/zhouhaifei-react-demo/src/routes.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
-import { BasicLayout } from './compnents/basicLayout';
+
+// import { BasicLayout } from './compnents/basicLayout';
 import { NotAuthority } from './pages/403';
 import { NotPage } from './pages/404';
 import type { MenuDataItem } from './utils';
 
 export const routes: MenuDataItem[] = [
   {
-    component: BasicLayout,
+    // component: BasicLayout,
     name: 'app',
     children: [
       {