+++ packages/zhouhaifei-react-scripts/config/entry.js
@@ -1,9 +1,3 @@
 const paths = require('./paths');
-const utils = require('./utils');
 
-module.exports = utils.isDevelopment ? [
-  require.resolve('../react-dev-utils/webpackHotDevClient'),
-
-  // Finally, this is your app's code:
-  paths.appIndexJs,
-] : { app: paths.appIndexJs };
+module.exports = { app: paths.appIndexJs };