+++ packages/zhouhaifei-react-scripts/react-dev-utils/checkRequiredFiles.js
@@ -1,17 +1,10 @@
-/**
- * Copyright (c) 2015-present, Facebook, Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
 'use strict';
 
 const fs = require('fs');
 const path = require('path');
 const chalk = require('chalk');
 
-function checkRequiredFiles(files) {
+module.exports = function checkRequiredFiles(files) {
   let currentFilePath;
   try {
     files.forEach((filePath) => {
@@ -27,6 +20,4 @@ function checkRequiredFiles(files) {
     console.log(chalk.red('  Searched in: ') + chalk.cyan(dirName));
     return false;
   }
-}
-
-module.exports = checkRequiredFiles;
+};