+++ packages/zhouhaifei-react-scripts/scripts/build.js
@@ -4,11 +4,6 @@
 process.env.BABEL_ENV = 'production';
 process.env.NODE_ENV = 'production';
 
-/*
- * Makes the script crash on unhandled rejections instead of silently
- * ignoring them. In the future, promise rejections that are not handled will
- * terminate the Node.js process with a non-zero exit code.
- */
 process.on('unhandledRejection', (err) => {
   throw err;
 });
@@ -23,7 +18,7 @@ const webpack = require('webpack');
 const paths = require('../config/paths');
 const utils = require('../config/utils');
 const configFactory = require('../config/webpack.config');
-const { checkBrowsers } = require('../react-dev-utils/browsersHelper');
+const checkBrowsers = require('../react-dev-utils/browsersHelper');
 const checkRequiredFiles = require('../react-dev-utils/checkRequiredFiles');
 const FileSizeReporter = require('../react-dev-utils/FileSizeReporter');
 const printHostingInstructions = require('../react-dev-utils/printHostingInstructions');
@@ -36,8 +31,6 @@ const useYarn = fs.existsSync(paths.yarnLockFile);
 const WARN_AFTER_BUNDLE_GZIP_SIZE = 512 * 1024;
 const WARN_AFTER_CHUNK_GZIP_SIZE = 1024 * 1024;
 
-const isInteractive = process.stdout.isTTY;
-
 // Warn and crash if required files are missing
 if (!checkRequiredFiles([paths.appIndexJs])) {
   process.exit(1);
@@ -46,71 +39,7 @@ if (!checkRequiredFiles([paths.appIndexJs])) {
 // Generate configuration
 const config = configFactory();
 
-/*
- * We require that you explicitly set browsers and do not fall back to
- * browserslist defaults.
- */
-
-checkBrowsers(paths.appPath, isInteractive).then(() => {
-  /*
-   * First, read the current file sizes in build directory.
-   * This lets us display how much they changed later.
-   */
-  return measureFileSizesBeforeBuild(paths.appDist);
-})
-  .then((previousFileSizes) => {
-    return build(previousFileSizes);
-  })
-  .then(
-    ({ stats, previousFileSizes }) => {
-      console.log(chalk.green('Compiled successfully.\n'));
-
-      console.log('File sizes after gzip:\n');
-      printFileSizesAfterBuild(
-        stats,
-        previousFileSizes,
-        paths.appDist,
-        WARN_AFTER_BUNDLE_GZIP_SIZE,
-        WARN_AFTER_CHUNK_GZIP_SIZE
-      );
-      console.log();
-
-      const appPackage = require(paths.appPackageJson);
-      const publicUrl = utils.publicUrlOrPath;
-      const publicPath = config.output.publicPath;
-      const buildFolder = path.relative(process.cwd(), paths.appDist);
-      printHostingInstructions(
-        appPackage,
-        publicUrl,
-        publicPath,
-        buildFolder,
-        useYarn
-      );
-    }
-  )
-  .catch((err) => {
-    console.log(chalk.red('Failed to compile.\n'));
-    process.exit(1);
-  });
-
-// Create the production build and print the deployment instructions.
-function build(previousFileSizes) {
-  /*
-   * We used to support resolving modules according to `NODE_PATH`.
-   * This now has been deprecated in favor of jsconfig/tsconfig.json
-   * This lets you use absolute paths in imports inside large monorepos:
-   */
-  if (process.env.NODE_PATH) {
-    console.log(
-      chalk.yellow(
-        'Setting NODE_PATH to resolve modules absolutely has been deprecated in favor of setting baseUrl in jsconfig.json (or tsconfig.json if you are using TypeScript) and will be removed in a future major release of create-react-app.'
-      )
-    );
-    console.log();
-  }
-
-  console.log('Creating an optimized production build...');
-
+function build() {
   const compiler = webpack(config);
   return new Promise((resolve, reject) => {
     compiler.run((err, stats) => {
@@ -121,24 +50,56 @@ function build(previousFileSizes) {
         }
 
         reject(new Error(err));
-        return;
-      }
+      } else {
+        const info = stats.toJson();
+        if (stats.hasErrors()) {
+          console.error(info.errors);
+          reject(new Error(info.errors));
+        } else {
+          if (stats.hasWarnings()) {
+            console.warn(info.warnings);
+          }
 
-      const info = stats.toJson();
-      if (stats.hasErrors()) {
-        console.error(info.errors);
-        reject(new Error(info.errors));
-        return;
+          resolve(stats);
+        }
       }
-
-      if (stats.hasWarnings()) {
-        console.warn(info.warnings);
-      }
-
-      return resolve({
-        stats,
-        previousFileSizes,
-      });
     });
   });
 }
+
+async function deploy() {
+  await checkBrowsers(paths.appPath);
+  const previousFileSizes = await measureFileSizesBeforeBuild(paths.appDist);
+  const stats = await build();
+
+  console.log(chalk.green('Compiled successfully.\n'));
+
+  console.log('File sizes after gzip:\n');
+  printFileSizesAfterBuild(
+    stats,
+    previousFileSizes,
+    paths.appDist,
+    WARN_AFTER_BUNDLE_GZIP_SIZE,
+    WARN_AFTER_CHUNK_GZIP_SIZE
+  );
+  console.log();
+
+  const appPackage = require(paths.appPackageJson);
+  const publicUrl = utils.publicUrlOrPath;
+  const publicPath = config.output.publicPath;
+  const buildFolder = path.relative(process.cwd(), paths.appDist);
+  printHostingInstructions(
+    appPackage,
+    publicUrl,
+    publicPath,
+    buildFolder,
+    useYarn
+  );
+}
+
+deploy().catch((err) => {
+  console.log(err);
+  console.log(chalk.red('Failed to compile.\n'));
+  process.exit(1);
+});
+